+++ src/main/java/com/codeup/adlister/models/User.java
@@ -3,7 +3,7 @@ package com.codeup.adlister.models;
 import com.codeup.adlister.util.Password;
 
 public class User {
-    private long id;
+    private int id;
     private String username;
     private String email;
     private String password;
@@ -16,18 +16,18 @@ public class User {
         setPassword(password);
     }
 
-    public User(long id, String username, String email, String password) {
+    public User(int id, String username, String email, String password) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.password = password;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
